requirements.txt
# Radio TUI Dependencies

# Core dependencies
requests>=2.28.0  # For HTTP requests to fetch now playing info

# Optional dependencies for YAML support
PyYAML>=6.0.0  # Optional: for YAML configuration support, falls back to JSON if unavailable
